+++ evm/evm-processor/src/ds-rpc/client.ts
@@ -1,5 +1,6 @@
 import {Logger} from '@subsquid/logger'
-import {RpcClient, RpcError} from '@subsquid/rpc-client'
+import {RetryError, RpcClient, RpcError} from '@subsquid/rpc-client'
+import {RpcRequest} from '@subsquid/rpc-client/lib/interfaces'
 import {assertNotNull, concurrentMap, def, groupBy, last, wait} from '@subsquid/util-internal'
 import {
     Batch,
@@ -179,7 +180,8 @@ export class EvmRpcDataSource implements HotDataSource<Block, DataRequest> {
             })
         }
         let blocks: rpc.Block[] = await this.rpc.batchCall(call, {
-            priority: s.range.from
+            priority: s.range.from,
+            validateResult: nonNull
         })
         for (let i = 1; i < blocks.length; i++) {
             assert.strictEqual(
@@ -202,15 +204,21 @@ export class EvmRpcDataSource implements HotDataSource<Block, DataRequest> {
         let subtasks = []
 
         if (req.logs && !req.receipts) {
-            subtasks.push(this.fetchLogs(blocks))
+            subtasks.push(catching(
+                this.fetchLogs(blocks)
+            ))
         }
 
         if (req.receipts) {
             let byBlockMethod = await this.getBlockReceiptsMethod()
             if (byBlockMethod) {
-                subtasks.push(this.fetchReceiptsByBlock(blocks, byBlockMethod))
+                subtasks.push(catching(
+                    this.fetchReceiptsByBlock(blocks, byBlockMethod)
+                ))
             } else {
-                subtasks.push(this.fetchReceiptsByTx(blocks))
+                subtasks.push(catching(
+                    this.fetchReceiptsByTx(blocks)
+                ))
             }
         }
 
@@ -328,7 +336,9 @@ export class EvmRpcDataSource implements HotDataSource<Block, DataRequest> {
 
         if (req.stateDiffs) {
             if (finalizedHeight < getBlockHeight(last(blocks)) || this.useDebugApiForStateDiffs) {
-                tasks.push(this.fetchDebugStateDiffs(blocks))
+                tasks.push(catching(
+                    this.fetchDebugStateDiffs(blocks)
+                ))
             } else {
                 replayTracers.push('stateDiff')
             }
@@ -337,17 +347,23 @@ export class EvmRpcDataSource implements HotDataSource<Block, DataRequest> {
         if (req.traces) {
             if (this.preferTraceApi) {
                 if (finalizedHeight < getBlockHeight(last(blocks)) || replayTracers.length == 0) {
-                    tasks.push(this.fetchTraceBlock(blocks))
+                    tasks.push(catching(
+                        this.fetchTraceBlock(blocks)
+                    ))
                 } else {
                     replayTracers.push('trace')
                 }
             } else {
-                tasks.push(this.fetchDebugFrames(blocks))
+                tasks.push(catching(
+                    this.fetchDebugFrames(blocks)
+                ))
             }
         }
 
         if (replayTracers.length) {
-            tasks.push(this.fetchReplays(blocks, replayTracers))
+            tasks.push(catching(
+                this.fetchReplays(blocks, replayTracers)
+            ))
         }
 
         return Promise.all(tasks).then()
@@ -531,4 +547,26 @@ function isConsistencyError(err: unknown): err is Error {
         if (/Expect block number from id/i.test(err.message)) return true
     }
     return false
+}
+
+
+function catching<T>(promise: Promise<T>): Promise<T> {
+    // prevent unhandled promise rejection crashes
+    promise.catch(() => {})
+    return promise
+}
+
+
+class UnexpectedResponse extends RetryError {
+    get name(): string {
+        return 'UnexpectedResponse'
+    }
+}
+
+
+function nonNull(result: any, req: RpcRequest): any {
+    if (result == null) throw new UnexpectedResponse(
+        `Result of call ${JSON.stringify(req)} was null. Perhaps, you should find a better endpoint.`
+    )
+    return result
 }