+++ evm/evm-processor/src/interfaces/chain.ts
@@ -1,6 +1,7 @@
 
 export interface RpcClient {
     call<T=any>(method: string, params?: unknown[]): Promise<T>
+    batchCall(batch: {method: string, params?: unknown[]}[]): Promise<any[]>
 }